JL8-103 refactor : Modal과 바텀 시트 위치 변경
